Code formatting - Remove duplicate code - Replace usage of var with explicit type declaration
