+++ tests/Benchmarks/Benchmarks.csproj
@@ -36,6 +36,7 @@
     <ProjectReference Include="..\..\src\System.Text.Http\System.Text.Http.csproj" />
     <ProjectReference Include="..\..\src\System.Text.JsonLab\System.Text.JsonLab.csproj" />
     <ProjectReference Include="..\..\src\System.Text.Utf8String\System.Text.Utf8String.csproj" />
+    <ProjectReference Include="..\System.Text.Primitives.Tests\System.Text.Primitives.Tests.csproj" />
   </ItemGroup>
   <ItemGroup>
     <Compile Update="JsonStrings.Designer.cs">
