+++ tests/Benchmarks/System.Text.Primitives/EncodeFromUtf8toUtf16.cs
@@ -1,7 +1,10 @@
-﻿using BenchmarkDotNet.Attributes;
+﻿// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
+using BenchmarkDotNet.Attributes;
 using System.Buffers;
 using System.Collections.Generic;
-using static System.Text.Primitives.Benchmarks.TextEncoderTestHelper;
 
 namespace System.Text.Primitives.Benchmarks
 {
@@ -9,7 +12,7 @@ namespace System.Text.Primitives.Benchmarks
     {
         public IEnumerable<CodePoint> GetEncodingPerformanceTestData()
         {
-            return TextEncoderTestHelper.GetEncodingPerformanceTestData();
+            return UtfEncoderHelper.GetEncodingPerformanceTestData();
         }
 
         [Params(99, 999, 9999)]
@@ -26,14 +29,14 @@ namespace System.Text.Primitives.Benchmarks
         [GlobalSetup]
         public void Setup()
         {
-            var inputString = GenerateStringData(Length, this.CodePointInfo.MinCodePoint, this.CodePointInfo.MaxCodePoint, this.CodePointInfo.Special);
+            string inputString = UtfEncoderHelper.GenerateStringData(Length, CodePointInfo.MinCodePoint, CodePointInfo.MaxCodePoint, CodePointInfo.Special);
             _characters = inputString.AsSpan().ToArray();
             _utf8Encoding = Encoding.UTF8;
             int utf8Length = _utf8Encoding.GetByteCount(_characters);
             _utf8Source = new byte[utf8Length];
             _utf8Encoding.GetBytes(_characters, 0, _characters.Length, _utf8Source, 0);
 
-            var status = Buffers.Text.Encodings.Utf8.ToUtf16Length(_utf8Source, out int needed);
+            OperationStatus status = Buffers.Text.Encodings.Utf8.ToUtf16Length(_utf8Source, out int needed);
             if (status != OperationStatus.Done)
                 throw new Exception();
 
@@ -49,7 +52,7 @@ namespace System.Text.Primitives.Benchmarks
         [Benchmark]
         public OperationStatus EncodeFromUtf8toUtf16UsingTextEncoder()
         {
-            var status = Buffers.Text.Encodings.Utf8.ToUtf16(_utf8Source, _utf16Destination, out int consumed, out int written);
+            OperationStatus status = Buffers.Text.Encodings.Utf8.ToUtf16(_utf8Source, _utf16Destination, out int consumed, out int written);
             if (status != OperationStatus.Done)
                 throw new Exception();
